Restructure to share code with method codec
