+++ shell/platform/common/cpp/client_wrapper/BUILD.gn
@@ -25,11 +25,14 @@ _wrapper_includes = [
 # reasonable (without forcing different kinds of clients to take unnecessary
 # code) to simplify use.
 _wrapper_sources = [
+  "byte_stream_wrappers.h",
   "engine_method_result.cc",
   "json_message_codec.cc",
   "json_method_codec.cc",
   "plugin_registrar.cc",
-  "standard_message_codec.cc",
+  "standard_codec_serializer.h",
+  "standard_codec.cc",
+  "standard_message_codec.cc",  # TODO fold into standard_codec.cc.
 ]
 
 # Client library build for internal use by the shell implementation.
@@ -79,6 +82,12 @@ executable("client_wrapper_unittests") {
     "method_call_unittests.cc",
     "plugin_registrar_unittests.cc",
     "standard_message_codec_unittests.cc",
+<<<<<<< HEAD
+=======
+    "standard_method_codec_unittests.cc",
+    "testing/encodable_value_utils.cc",
+    "testing/encodable_value_utils.h",
+>>>>>>> e547fa656... Restructure to share code with method codec
   ]
 
   deps = [
