+++ shell/platform/common/cpp/client_wrapper/include/flutter/standard_message_codec.h
@@ -10,9 +10,6 @@
 
 namespace flutter {
 
-class StandardCodecByteStreamReader;
-class StandardCodecByteStreamWriter;
-
 // A binary message encoding/decoding mechanism for communications to/from the
 // Flutter engine via message channels.
 class StandardMessageCodec : public MessageCodec<EncodableValue> {
@@ -38,33 +35,6 @@ class StandardMessageCodec : public MessageCodec<EncodableValue> {
   // |flutter::MessageCodec|
   std::unique_ptr<std::vector<uint8_t>> EncodeMessageInternal(
       const EncodableValue& message) const override;
-
- private:
-  // Reads and returns then next value from |stream|.
-  EncodableValue ReadValue(StandardCodecByteStreamReader* stream) const;
-
-  // Writes the encoding of |value| to |stream|.
-  void WriteValue(const EncodableValue& value,
-                  StandardCodecByteStreamWriter* stream) const;
-
-  // Reads the variable-length size from the current position in |stream|.
-  uint32_t ReadSize(StandardCodecByteStreamReader* stream) const;
-
-  // Writes the variable-length size encoding to |stream|.
-  void WriteSize(uint32_t size, StandardCodecByteStreamWriter* stream) const;
-
-  // Reads a fixed-type list whose values are of type T from the current
-  // position in |stream|, and returns it as the corresponding EncodableValue.
-  // |T| must correspond to one of the support list value types of
-  // EncodableValue.
-  template <typename T>
-  EncodableValue ReadVector(StandardCodecByteStreamReader* stream) const;
-
-  // Writes |vector| to |stream| as a fixed-type list. |T| must correspond to
-  // one of the support list value types of EncodableValue.
-  template <typename T>
-  void WriteVector(const std::vector<T> vector,
-                   StandardCodecByteStreamWriter* stream) const;
 };
 
 }  // namespace flutter
